refactor: Move deposit logic to deposit.js and fix minor issues
- Moved deposit logic to deposit.js for better code organization
- Fixed minor issues related to user input validation
